third_party/blink/web_tests/http/tests/websocket/workers/receive-arraybuffer-expected.txt

WebSocket: Receive ArrayBuffers in Web Workers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS PASS: ws.binaryType is "arraybuffer"
PASS PASS: receivedMessages.length is 3
INFO: Checking message #0.
PASS PASS: responseType is "[object ArrayBuffer]"
PASS PASS: actualArray.length is 13
PASS PASS: Passed: Message #0.
INFO: Checking message #1.
PASS PASS: responseType is "[object ArrayBuffer]"
PASS PASS: actualArray.length is 0
PASS PASS: Passed: Message #1.
INFO: Checking message #2.
PASS PASS: responseType is "[object ArrayBuffer]"
PASS PASS: actualArray.length is 256
PASS PASS: Passed: Message #2.
DONE
PASS successfullyParsed is true

TEST COMPLETE
